[MAIN GROUP 8][REFACTOR][BACK END][Implementa nova forma de tratar erros]
