+++ back-end/src/api/middlewares/errorHandler.js
@@ -3,18 +3,20 @@ const httpStatusCode = require('../utils/httStatusCodes');
 const errorHandler = (err, _req, res, _next) => {
   if (err.isJoi) {
     const { details } = err;
-    return res.status(httpStatusCode.badRequest).json({ message: details[0].message });
+    return res.status(httpStatusCode.badRequest)
+    .json({ message: details[0].message, code: 'badRequest' });
   }
 
   if (err.code) {
     const { code, message } = err;
-    return res.status(httpStatusCode[code]).json({ message });
+    return res.status(httpStatusCode[code]).json({ message, code });
   }
 
   console.log(err);
 
   return res.status(httpStatusCode.internalServerError).json({
-      message: 'Oooops D: internal Server Error', 
+      message: 'Oooops D: internal Server Error',
+      code: 'internalServerError',
   });
 };
