+++ back-end/src/api/middlewares/tokenGenerator.js
