+++ back-end/src/api/services/loginService.js
@@ -1,7 +1,8 @@
 const { createHash } = require('crypto');
 const { User } = require('../../database/models');
 
-const generateToken = require('../middlewares/generateToken');
+const generateToken = require('../middlewares/tokenGenerator');
+const errorTypes = require('../utils/errorTypes');
 
 const login = async (email, password) => {
   const cryptoPassword = createHash('md5').update(password).digest('hex');
@@ -9,17 +10,16 @@ const login = async (email, password) => {
   const foundUserData = await User.findOne({ where: { email, password: cryptoPassword } });
 
   if (!foundUserData) {
-    throw Object.assign(
-      new Error('Invalid email or password'),
-      { code: 'notFound' },
-   );
+    const error = errorTypes.invalidCredentials;
+
+    return { error };
   }
   
   const { password: _, ...userData } = foundUserData.dataValues;
 
   const token = await generateToken(userData);
 
-  return token;
+  return { token };
 };
 
 module.exports = { login };
