+++ back-end/src/api/utils/errorTypes.js
@@ -1,0 +1,6 @@
+module.exports = {
+  invalidCredentials: {
+    message: 'Invalid email or password',
+    code: 'notFound',
+  },
+};
